// ui/src/api/generated/model/plantListResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * APIs for Whirlpool Inspection Tool API developed by Scopt Analytics for Whirlpool Inspection Tool API platform.
 * OpenAPI spec version: 1.0.3
 */
import type { PlantResponse } from './plantResponse';

export interface PlantListResponse {
  data: PlantResponse[];
  total: number;
  page: number;
  per_page: number;
  total_pages: number;
}